+++ starsky/starsky/Controllers/MemoryCacheDebugController.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Text.Json;
 using System.Text.Json.Serialization;
@@ -5,6 +6,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Caching.Memory;
 using starsky.foundation.platform.Extensions;
+using starsky.foundation.platform.Interfaces;
 
 namespace starsky.Controllers
 {
@@ -12,10 +14,12 @@ namespace starsky.Controllers
 	public class MemoryCacheDebugController: Controller
 	{
 		private readonly IMemoryCache _memoryCache;
+		private readonly IWebLogger _logger;
 
-		public MemoryCacheDebugController(IMemoryCache memoryCache)
+		public MemoryCacheDebugController(IMemoryCache memoryCache, IWebLogger logger)
 		{
 			_memoryCache = memoryCache;
+			_logger = logger;
 		}
 		
 		[HttpGet("/api/MemoryCacheDebug")]
@@ -25,7 +29,15 @@ namespace starsky.Controllers
 			foreach ( var key in _memoryCache.GetKeys<string>() )
 			{
 				_memoryCache.TryGetValue(key, out var data);
-				result.Add(key, JsonSerializer.Serialize(data));
+				try
+				{
+					result.Add(key, JsonSerializer.Serialize(data));
+				}
+				catch ( JsonException exception )
+				{
+					_logger.LogError(exception, $"[MemoryCacheDebug] Json {key} has failed (catch-ed)");
+					result.Add(key, "[ERROR] Has failed parsing");
+				}
 			}
 			return Json(result);
 		}