+++ starskyapp/src/app/child-process/setup-child-process.ts
@@ -31,15 +31,20 @@ export async function setupChildProcess() {
 
   appPort = await getFreePort();
 
+  logger.info('-appSettingsPath >');
+  logger.info(appSettingsPath);
+
   const env = {
     ASPNETCORE_URLS: `http://localhost:${appPort}`,
     app__thumbnailTempFolder: thumbnailTempFolder,
     app__tempFolder: tempFolder,
-    app__appSettingsPath: appSettingsPath,
+    app__appsettingspath: appSettingsPath,
+    app__NoAccountLocalhost: "true",
     app__databaseConnection: databaseConnection,
     app__AccountRegisterDefaultRole: "Administrator",
     app__Verbose: !isPackaged() ? "true" : "false"
   };
+  process.env = {...process.env, ...env};
 
   logger.info("env settings ->");
   logger.info(env);
@@ -55,7 +60,7 @@ export async function setupChildProcess() {
   const starskyChild = spawn(appStarskyPath, {
     cwd: path.dirname(appStarskyPath),
     detached: true,
-    env
+    env: process.env,
   });
 
   starskyChild.stdout.on("data", function (data) {